requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pydantic==2.5.0
click==8.1.7
python-dotenv==1.0.0
streamlit==1.28.1
pandas==1.5.3
openai==1.3.0
requests==2.31.0
paramiko==3.3.1

# OCR Dependencies for scanned PDF/image support
paddleocr==3.0.0
paddlepaddle==3.1.0
pillow==10.1.0
reportlab==4.0.7
pymupdf==1.23.8